Refatoração Controller, Business e Router
