+++ modulo7/projeto-cookenu/src/app.ts
@@ -1,6 +1,8 @@
 import express from "express";
 import cors from 'cors'
-import router from "./routes/routes";
+import userRouter from "./routes/UserRouter";
+import friendsRouter from "./routes/FriendsRouter";
+import recipesRouter from "./routes/RecipesRouter";
 
 
 
@@ -11,6 +13,8 @@ const app = express()
 app.use(express.json())
 app.use(cors())
 
-app.use("/user" , router)
+app.use("/user" , userRouter)
+app.use("/friends" , friendsRouter )
+app.use("/recipes" , recipesRouter )
 
 export default app
