+++ modulo7/projeto-cookenu/src/routes/FriendsRouter.ts
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import { FriendsCT } from "../controller/FriendsController";
+
+
+ const friendsRouter = Router()
+const friendsCT = new FriendsCT()
+
+
+friendsRouter.post("/follow" , friendsCT.makeFriends)
+friendsRouter.delete("/unfloow" , friendsCT.dontFriend)
+
+export default friendsRouter
