+++ modulo7/projeto-cookenu/src/routes/RecipesRouter.ts
@@ -1,0 +1,12 @@
+import { Router } from "express";
+import { RecipesCT } from '../controller/RecipesController'
+
+
+const recipesRouter = Router()
+const recipesCT = new RecipesCT()
+
+recipesRouter.post("/create" , recipesCT.createRecipes)
+recipesRouter.get("/:id" , recipesCT.informationRecipes)
+
+
+export default recipesRouter
